+++ src/app/common/authentication.service.ts
@@ -34,7 +34,7 @@ export class AuthenticationService {
           if (user && user.identity.token) {
             // store user details and jwt token in local storage to keep user logged in between page refreshes
             localStorage.setItem("currentUser", JSON.stringify(user.identity));
-            this.currentUserSubject.next(user);
+            this.currentUserSubject.next(user.identity);
           }
 
           return user;