+++ src/server/Login/LoginRoute.ts
@@ -13,4 +13,16 @@ router.post("/authenticate", async (req, res) => {
   server.login(node);
 });
 
+/**
+ * Validates token
+ */
+router.get("/validate", async (req, res) => {
+  let server: LoginServer = new LoginServer();
+  let node: iNode = {
+    request: req,
+    response: res
+  };
+  server.validateToken(node);
+});
+
 export { router };