+++ src/server/SecretForToken.ts
@@ -1,0 +1,5 @@
+import { configModule } from "./ConfigModule";
+
+export const secretForToken: string = configModule.getConfigValue(
+  "secretForToken"
+);